# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recipe Book API",
        "description": "API for managing recipes, cuisines, ingredients, and user authentication",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/attachments": {
            "post": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Upload a file (image)",
                "description": "Upload a file",
                "operationId": "514387e0f7b86358ce45fd37883514fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Image file (max 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost/storage/attachments/image.jpg"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "image.jpg"
                                        },
                                        "original_name": {
                                            "type": "string",
                                            "example": "photo.jpg"
                                        },
                                        "mime": {
                                            "type": "string",
                                            "example": "image/jpeg"
                                        },
                                        "size": {
                                            "type": "integer",
                                            "example": 102400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            }
        },
        "/api/attachments/{id}": {
            "get": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Get attachment by ID",
                "description": "Get attachment by ID",
                "operationId": "70172d772dc02144178f2ce0e1556940",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attachment details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost/storage/attachments/image.jpg"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "image.jpg"
                                        },
                                        "original_name": {
                                            "type": "string",
                                            "example": "photo.jpg"
                                        },
                                        "mime": {
                                            "type": "string",
                                            "example": "image/jpeg"
                                        },
                                        "size": {
                                            "type": "integer",
                                            "example": 102400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "session": {
                "type": "apiKey",
                "description": "Session-based authentication using Laravel sessions",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}